backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
numpy
pymongo
requests
python-dotenv
